Extract anonymous marking summary and editor elements
